All Project

+++ backend/controllers/BlogsController.php
@@ -213,6 +213,9 @@ class BlogsController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
 
         $tagQuery = new Query;
 
@@ -331,6 +334,10 @@ class BlogsController extends Controller
      */
     public function actionDelete($id)
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
+        
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
